+++ running/databasemanagerrunning.py
@@ -16,11 +16,11 @@ class DatabaseManagerRunning(sqlite3.Connection):
         def factory(row):
             date = datetime.date.fromordinal(row[1]).isoformat().split("-")
             year, month, day = map(int, date)
-            return SessionRunning(year, month, day, *row[2:])
+            return (row[0], SessionRunning(year, month, day, *row[2:]))
         self.row_factory = lambda cursor, row: factory(row)
         self.c = self.cursor()
-        self.row_factory = lambda cursor, row: row[0]
-        self.cl = self.cursor() # list cursor (returns list, use when only one field is requested)
+        #self.row_factory = lambda cursor, row: row[0]
+        #self.cl = self.cursor() # list cursor (returns list, use when only one field is requested)
         self.row_factory = None
         # Attributes
         self.database = database
@@ -75,24 +75,36 @@ class DatabaseManagerRunning(sqlite3.Connection):
         if commit: self.commit()
     
     
-    def selectone(self, year: int, month: int, day: int):
+    def select_one(self, idsession: int, check:bool =False):
         """
-        Select one 'SessionRunning' from the database given its date
+        Select one 'SessionRunning' from the database given its idsession
         """
+        
+        if check: self.check_if_id_in_database_id(idsession)
+        query = "SELECT * FROM {} WHERE idsession = {}".format(self.tablename, idsession)
+        self.c.execute(query)
+        return self.c.fetchone()
+    
+    
+    def select_date(self, year: int, month: int, day: int, check: bool=False):
+        """
+        Select 'SessionRunning's from the database given its date
+        """
+        
         # Check if the 'SessionRunning' is registered on the database
         date = datetime.date(year, month, day).toordinal()
-        self.check_if_in_database(date)
+        if check: self.check_if_date_in_database(date)
         # Select
         query = "SELECT * FROM {} WHERE date = {}".format(
             self.tablename,
             date
             )
         self.c.execute(query)
-        sessionRunning = self.c.fetchone()
+        sessionRunning = self.c.fetchall()
         return sessionRunning
     
     
-    def selectmulti(self, year0: int, month0: int, day0: int, year1: int,
+    def select_multi_date(self, year0: int, month0: int, day0: int, year1: int,
         month1: int, day1: int):
         """
         Select all the 'SessionRunning's from the database given the limits of
@@ -108,7 +120,7 @@ class DatabaseManagerRunning(sqlite3.Connection):
             yield self.c.fetchone()
     
     
-    def selectall(self):
+    def select_all(self):
         """
         Select all 'SessionRunning's from the database
         """
@@ -119,42 +131,28 @@ class DatabaseManagerRunning(sqlite3.Connection):
             yield self.c.fetchone()
     
     
-    def update(self, col: str, value, year0: int, month0: int, day0: int, 
-        commit: bool = True):
-        """
-        Update the data of a 'SessionRunning' given its date
-        """
-        # The idsession is unique, so it cannot be changed to avoid errors
-        if col == "idsession":
-            msg = "The idsession cannot be changed"
-            raise DataBaseManagerRunningException(msg)
-        # Check if the 'SessionRunning' is registered on the database
-        date0 = datetime.date(year0, month0, day0).toordinal()
-        self.check_if_in_database(date0)
-        # If the date is requested to be updated, some changes are needed
-        if col == "date":
-            year1, month1, day1 = map(int, value.split("-"))
-            date1 = datetime.date(year1, month1, day1).toordinal()
-            value = date1
-        # Update
-        query = "UPDATE {} SET {} = {} WHERE date = {}".format(
-            self.tablename,
-            col,
-            value,
-            date0
-            )
-        self.c.execute(query)
-        if commit: self.commit()
-    
-    
-    def update_all(self, idsession, year, month, day, distance, time, kcal, commit = True):
+    def update(self, idsession, year: int = None, month: int = None,
+        day: int = None, distance: int = None, time: int = None,
+        kcal: int = None, check: bool = False, commit:bool = True):
         """
         Update all the data of a 'SessionRunning' given its idsession
         """
         # Check if the 'SessionRunning' is registered on the database
         
+        if check: self.check_if_id_in_database(idsession)
         
-        sessionRunning = SessionRunning(year, month, day, distance, time, kcal)
+        currSessionRunning = self.select_one(idsession)[1]
+        currDate = currSessionRunning.date_dict()
+        
+        if year == None: year = currDate["year"]
+        if month == None: month = currDate["month"]
+        if day == None: day = currDate["day"]
+        if distance == None: distance = currSessionRunning.distance
+        if time == None: time = currSessionRunning.time
+        if kcal == None: kcal = currSessionRunning.kcal
+        
+        newSessionRunning = SessionRunning(year, month, day, distance, time, kcal)
+        
         # Update
         query = ("UPDATE {} SET "
             "date = {}, "
@@ -163,7 +161,7 @@ class DatabaseManagerRunning(sqlite3.Connection):
             "kcal = {} "
             "WHERE idsession = {}").format(
             self.tablename,
-            sessionRunning.date, distance, time, kcal, idsession
+            newSessionRunning.date, distance, time, kcal, idsession
             )
         self.c.execute(query)
         if commit: self.commit()
@@ -182,7 +180,7 @@ class DatabaseManagerRunning(sqlite3.Connection):
         if commit: self.commit()
     
     
-    def check_if_in_database(self, date: int):
+    def check_if_date_in_database(self, date: int):
         """
         Checks if there is a 'SessionRunning' with the given date in the database
         """
@@ -191,6 +189,18 @@ class DatabaseManagerRunning(sqlite3.Connection):
         data = self.c.fetchone()
         if data is None:
             msg = "There is no record with that date"
+            raise DataBaseManagerRunningException(msg)
+    
+    
+    def check_if_id_in_database(self, idsession: int):
+        """
+        Checks if there is a 'SessionRunning' with the given idsession
+        """
+        query = "SELECT * FROM {} WHERE idsession = {}".format(self.tablename, idsession)
+        self.c.execute(query)
+        data = self.c.fetchone()
+        if data is None:
+            msg = "There is no record with that idsession"
             raise DataBaseManagerRunningException(msg)